requirements.txt
Flask>=2.2
Flask-PyMongo>=2.3
python-dotenv>=1.0
requests>=2.28
dnspython>=2.0